Refactor multi-vendor architecture to utilize a delegate pattern for BLE connection and synchronization lifecycles.
- Introduced `VendorConnectionDelegate` and `DefaultConnectionDelegate` to encapsulate vendor-specific connection logic, such as MTU negotiation, handshakes, and retry strategies.
- Implemented `SonyConnectionDelegate` to handle complex Sony-specific requirements, including DD30/DD31 session locking, notification subscriptions, and capability-based location packet encoding.
- Refactored `KableCameraRepository` and `KableCameraConnection` to delegate sync operations (`syncLocation`, `syncDateTime`) and connection setup to the new delegate architecture, significantly simplifying core repository logic.
- Updated the `Camera` model and `CameraVendor` interface to support generic `vendorMetadata`, replacing hardcoded Sony protocol version fields with a flexible map.
- Removed deprecated backward compatibility type aliases and classes for Ricoh protocol and GATT specs.
- Updated `MULTI_VENDOR_SUPPORT.md` and `AGENTS.md` documentation to reflect the new architecture and delegate-based extension flow.

These changes improve code maintainability and isolation by keeping vendor-specific complexities out of the core domain and repository layers.
