+++ app/src/main/kotlin/dev/sebastiano/camerasync/domain/model/Camera.kt
@@ -14,11 +14,10 @@ data class Camera(
     val macAddress: String,
     val vendor: CameraVendor,
     /**
-     * BLE protocol version from the advertisement data. For Sony cameras, this determines which
-     * features are available:
-     * - Protocol >= 65: Requires DD30/DD31 unlock sequence for location sync
-     * - Protocol < 65: Uses legacy protocol without DD30/DD31 Null for non-Sony cameras or if not
-     *   available in the advertisement.
+     * Vendor-specific metadata parsed from the advertisement.
+     *
+     * This generic map allows vendors to store discovery-time information (like protocol versions)
+     * without polluting the core domain model with vendor-specific fields.
      */
-    val bleProtocolVersion: Int? = null,
+    val vendorMetadata: Map<String, Any> = emptyMap(),
 )
