+++ app/src/main/kotlin/dev/sebastiano/camerasync/scanning/ScanningViewModel.kt
@@ -16,7 +16,6 @@ import dev.sebastiano.camerasync.ble.buildManufacturerDataMap
 import dev.sebastiano.camerasync.domain.model.Camera
 import dev.sebastiano.camerasync.domain.vendor.CameraVendorRegistry
 import dev.sebastiano.camerasync.logging.KhronicleLogEngine
-import dev.sebastiano.camerasync.vendors.sony.SonyCameraVendor
 import kotlin.uuid.ExperimentalUuidApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -126,13 +125,8 @@ internal class ScanningViewModel(private val vendorRegistry: CameraVendorRegistr
             return null
         }
 
-        // Parse BLE protocol version for Sony cameras
-        val protocolVersion =
-            if (vendor.vendorId == "sony") {
-                SonyCameraVendor.parseProtocolVersion(mfrData)
-            } else {
-                null
-            }
+        // Generic metadata parsing
+        val metadata = vendor.parseAdvertisementMetadata(mfrData)
 
         Log.info(tag = TAG) { "Discovered ${vendor.vendorName} camera: ${peripheralName ?: name}" }
         return Camera(
@@ -140,7 +134,7 @@ internal class ScanningViewModel(private val vendorRegistry: CameraVendorRegistr
             name = peripheralName ?: name,
             macAddress = identifier,
             vendor = vendor,
-            bleProtocolVersion = protocolVersion,
+            vendorMetadata = metadata,
         )
     }
 }
